rename  locale.py for app_locale.py to avoid confusion.
